LUN-306: Move AlternativeWordsContainer to own component
