+++ src/navigation/Navigator.tsx
@@ -7,12 +7,12 @@ import labels from '../constants/labels.json'
 import useReadSelectedProfessions from '../hooks/useReadSelectedProfessions'
 import { useTabletHeaderHeight } from '../hooks/useTabletHeaderHeight'
 import ProfessionSelectionScreen from '../routes/ProfessionSelectionScreen'
-import VocabularyDetailScreen from '../routes/VocabularyDetailScreen'
 import VocabularyListScreen from '../routes/VocabularyListScreen'
 import ArticleChoiceExerciseScreen from '../routes/choice-exercises/ArticleChoiceExerciseScreen'
 import WordChoiceExerciseScreen from '../routes/choice-exercises/WordChoiceExerciseScreen'
 import ExerciseFinishedScreen from '../routes/exercise-finished/ExerciseFinishedScreen'
 import ScopeSelection from '../routes/scope-selection/ScopeSelectionScreen'
+import VocabularyDetailScreen from '../routes/vocabulary-detail/VocabularyDetailScreen'
 import WriteExerciseScreen from '../routes/write-exercise/WriteExerciseScreen'
 import BottomTabNavigator from './BottomTabNavigator'
 import { RoutesParams } from './NavigationTypes'
